test(pinia-colada): colocate streamed & live query tests with source
Move the real-Pinia-Colada tests for serializableStreamedQuery and liveQuery
from tests/ back next to their source files as src/*.test.tsx, with the
mount and chunk-stream helpers inlined per file instead of imported from a
shared module.
